Fixed issue where feedback was not present when editing bid failed.
Prior to this update, if you attempted to edit your bid when the auction was two days  away or sooner, it would appear like you actually accomplished the action, but in actuality this was not allowed and no change was really made. Now, when you attempt to edit a bid of this nature, it will instead prompt the user that doing so is impossible, and as such, the program will not allow it.